Working on Tagging System
Currently pretty messy code-wise, I wrote tests for one of the two new DAOs (CreatedTagsDAO). The other one I will have to do at a later date.

You can click on a tag to remove it, and click on the plus to add a new tag. When the popup opens, you will have the option to select from your own pool of tags, or create new ones. Can't currently edit the colours or names.