User login and register changes
Username and password are now checked to ensure they meet the quality requirements.

Username must be between 4 and 20 chars
Password must be at least 8 characters
password must contain letters, a number and a special character.

Passwords must also now match with the repeated password field.

On the login screen, the user is now informed if the account doesn't exist, and if it does, they are told if their password is incorrect.

Updated loading screen to show up if the task is taking a "long" time, instead of showing up every time.